Creating exception handlers for create operations

+++ creep-tenuous-core/src/main/java/com/zer0s2m/creeptenuous/core/handlers/impl/ServiceFileSystemExceptionHandlerOperationCreateImpl.java
@@ -2,6 +2,7 @@ package com.zer0s2m.creeptenuous.core.handlers.impl;
 
 import com.zer0s2m.creeptenuous.core.annotations.AtomicFileSystemExceptionHandler;
 import com.zer0s2m.creeptenuous.core.context.ContextAtomicFileSystem;
+import com.zer0s2m.creeptenuous.core.context.ContextAtomicFileSystem.Operations;
 import com.zer0s2m.creeptenuous.core.handlers.ServiceFileSystemExceptionHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -13,11 +14,12 @@ import java.nio.file.Path;
 
 /**
  * The base interface for handling exceptions that interact with the file system.
+ * Operation type handler {@link ContextAtomicFileSystem.Operations#CREATE}
  * <p>Used in conjunction with an annotation: {@link AtomicFileSystemExceptionHandler#handler()}</p>
  */
-public class ServiceFileSystemExceptionHandlerImpl implements ServiceFileSystemExceptionHandler {
+public class ServiceFileSystemExceptionHandlerOperationCreateImpl implements ServiceFileSystemExceptionHandler {
 
-    private final Logger logger = LoggerFactory.getLogger(ServiceFileSystemExceptionHandlerImpl.class);
+    private final Logger logger = LoggerFactory.getLogger(ServiceFileSystemExceptionHandlerOperationCreateImpl.class);
 
     /**
      * Context for working with the file system in <b>atomic mode</b>
@@ -25,7 +27,8 @@ public class ServiceFileSystemExceptionHandlerImpl implements ServiceFileSystemE
     private final ContextAtomicFileSystem contextAtomicFileSystem = ContextAtomicFileSystem.getInstance();
 
     /**
-     * Handle exception caused by file system.
+     * Handling an exception thrown by the file system. Each handler must be
+     * responsible for one type of operation {@link ContextAtomicFileSystem.Operations}
      * <p>After handling the exception, call atomic mode handling
      * {@link com.zer0s2m.creeptenuous.core.context.ContextAtomicFileSystem#handleOperation(ContextAtomicFileSystem.Operations, String)}
      * to clear the context</p>
@@ -36,20 +39,23 @@ public class ServiceFileSystemExceptionHandlerImpl implements ServiceFileSystemE
     @Override
     public void handleException(Throwable t, HashMap<String, HashMap<String, Object>> operationsData) {
         operationsData.forEach((uniqueName, operationData) -> {
-            Path systemPath = (Path) operationData.get("systemPath");
-            try {
-                logger.info(String.format(
-                        "Delete file atomic mode [%s]: [%s]",
-                        systemPath, Files.deleteIfExists(systemPath)
-                ));
-            } catch (IOException e) {
-                throw new RuntimeException(e);
-            }
+            Operations typeOperation = (Operations) operationData.get("operation");
+            if (typeOperation.equals(Operations.CREATE)) {
+                Path systemPath = (Path) operationData.get("systemPath");
+                try {
+                    logger.info(String.format(
+                            "Delete file atomic mode [%s]: [%s]",
+                            systemPath, Files.deleteIfExists(systemPath)
+                    ));
+                } catch (IOException e) {
+                    throw new RuntimeException(e);
+                }
 
-            contextAtomicFileSystem.handleOperation(
-                    (ContextAtomicFileSystem.Operations) operationData.get("operation"),
-                    uniqueName
-            );
+                contextAtomicFileSystem.handleOperation(
+                        (ContextAtomicFileSystem.Operations) operationData.get("operation"),
+                        uniqueName
+                );
+            }
         });
     }
 }
